refactor/requirements.txt
# Core Web Framework
fastapi==0.116.0
uvicorn==0.35.0
python-multipart==0.0.20

# HTTP Client
httpx==0.28.1
aiohttp==3.12.13

# Utilities
tenacity==9.1.2
openai==1.93.1
pydantic==2.11.7
python-dotenv==1.1.1a